Split login window code out of navigation file.
